Move ModelPrior to model extensions (#355)
* Initial commit with CHANGELOG.rst.

* Move ModelPrior from methods to model.

* Change ModelPrior import location.

* Refactor ModelPrior from model.util to model.extensions.

* Update CHANGELOG.rst
